查询平台信息免锁和直接返回u16 groups列表
**复现步骤**

**原因和解决方案**

**影响范围**

**验证方案**

**涉及分支**

* master
* 5.7.1

**检查项**

- [ ] 需要更新依赖
- [ ] 是共性问题(代码中存在类似问题)
- [ ] 编译通过
- [ ] 单元测试通过

/assign @jinjie

/cc @yuanchao @xiangyang

+++ reciter-api/request.go
@@ -128,7 +128,7 @@ const (
 
 type GroupIDMap struct {
 	L3EpcID     int32    `json:"l3_epc_id"`
-	GroupID     int32    `json:"group_id"`
+	GroupID     uint16   `json:"group_id"`
 	CIDRs       []string `json:"cidrs"`
 	IPRanges    []string `json:"ip_ranges"` // 格式为{IP0}-{IP1}, 如 10.33.2.202-10.34.1.233
 	PodGroupID  uint16   `json:"pod_group_id"`
